ENT-2294 | Moving utility functions in program enrollments to places related to the data they manipulate
