+++ lms/djangoapps/bulk_email/api.py
@@ -11,3 +11,20 @@ from bulk_email.models_api import (
     is_bulk_email_feature_enabled,
     is_user_opted_out_for_course
 )
+
+
+def get_emails_enabled(user, course_id):
+    """
+    Get whether or not emails are enabled in the context of a course.
+
+    Arguments:
+        user: the user object for which we want to check whether emails are enabled
+        course_id (string): the course id of the course
+
+    Returns:
+        (bool): True if emails are enabled for the course associated with course_id for the user;
+        False otherwise
+    """
+    if is_bulk_email_feature_enabled(course_id=course_id):
+        return not is_user_opted_out_for_course(user=user, course_id=course_id)
+    return None
