+++ lms/djangoapps/course_api/api.py
@@ -4,8 +4,10 @@ Course API
 
 from __future__ import absolute_import
 
+from edx_when.api import get_dates_for_course
 from django.conf import settings
 from django.contrib.auth.models import AnonymousUser, User
+from django.urls import reverse
 from rest_framework.exceptions import PermissionDenied
 import search
 import six
@@ -17,6 +19,7 @@ from lms.djangoapps.courseware.courses import (
     get_permission_for_course_about
 )
 from openedx.core.lib.api.view_utils import LazySequence
+from xmodule.modulestore.django import modulestore
 
 from .permissions import can_view_courses_for_username
 
@@ -148,3 +151,58 @@ def list_courses(request, username, org=None, roles=None, filter_=None, search_t
     course_qs = _filter_by_role(course_qs, user, roles)
     course_qs = _filter_by_search(course_qs, search_term)
     return course_qs
+
+
+def get_due_dates(request, course_key, user):
+    """
+    Get due date information for a user for blocks in a course.
+
+    Arguments:
+        request: the request object
+        course_key (CourseKey): the CourseKey for the course
+        user: the user object for which we want due date information
+
+    Returns:
+        due_dates (list): a list of dictionaries containing due date information
+            keys:
+                name: the display name of the block
+                url: the deep link to the block
+                date: the due date for the block
+    """
+    dates = get_dates_for_course(
+        course_key,
+        user,
+    )
+
+    store = modulestore()
+
+    due_dates = []
+    for (block_key, date_type), date in six.iteritems(dates):
+        if date_type == 'due':
+            block = store.get_item(block_key)
+
+            # get url to the block in the course
+            block_url = reverse('jump_to', args=[course_key, block_key])
+            block_url = request.build_absolute_uri(block_url)
+
+            due_dates.append({
+                'name': block.display_name,
+                'url': block_url,
+                'date': date,
+            })
+    return due_dates
+
+
+def get_course_run_url(request, course_id):
+    """
+    Get the URL to a course run.
+
+    Arguments:
+        request: the request object
+        course_id (string): the course id of the course
+
+    Returns:
+        (string): the URL to the course run associated with course_id
+    """
+    course_run_url = reverse('openedx.course_experience.course_home', args=[course_id])
+    return request.build_absolute_uri(course_run_url)
