+++ lms/djangoapps/program_enrollments/rest_api/v1/utils.py
@@ -7,19 +7,15 @@ from __future__ import absolute_import, unicode_literals
 from datetime import datetime, timedelta
 from functools import wraps
 
-from django.urls import reverse
 from django.utils.functional import cached_property
-from edx_when.api import get_dates_for_course
 from opaque_keys.edx.keys import CourseKey
 from pytz import UTC
 from rest_framework import status
-from six import iteritems
 
 from bulk_email.api import is_bulk_email_feature_enabled, is_user_opted_out_for_course
 from lms.djangoapps.grades.rest_api.v1.utils import CourseEnrollmentPagination
 from openedx.core.djangoapps.catalog.utils import get_programs, is_course_run_in_program
 from openedx.core.lib.api.view_utils import verify_course_exists
-from xmodule.modulestore.django import modulestore
 
 from .constants import CourseRunProgressStatuses
 
@@ -118,78 +114,6 @@ def verify_course_exists_and_in_program(view_func):
             )
         return view_func(self, request, **kwargs)
     return wrapped_function
-
-
-def get_due_dates(request, course_key, user):
-    """
-    Get due date information for a user for blocks in a course.
-
-    Arguments:
-        request: the request object
-        course_key (CourseKey): the CourseKey for the course
-        user: the user object for which we want due date information
-
-    Returns:
-        due_dates (list): a list of dictionaries containing due date information
-            keys:
-                name: the display name of the block
-                url: the deep link to the block
-                date: the due date for the block
-    """
-    dates = get_dates_for_course(
-        course_key,
-        user,
-    )
-
-    store = modulestore()
-
-    due_dates = []
-    for (block_key, date_type), date in iteritems(dates):
-        if date_type == 'due':
-            block = store.get_item(block_key)
-
-            # get url to the block in the course
-            block_url = reverse('jump_to', args=[course_key, block_key])
-            block_url = request.build_absolute_uri(block_url)
-
-            due_dates.append({
-                'name': block.display_name,
-                'url': block_url,
-                'date': date,
-            })
-    return due_dates
-
-
-def get_course_run_url(request, course_id):
-    """
-    Get the URL to a course run.
-
-    Arguments:
-        request: the request object
-        course_id (string): the course id of the course
-
-    Returns:
-        (string): the URL to the course run associated with course_id
-    """
-    course_run_url = reverse('openedx.course_experience.course_home', args=[course_id])
-    return request.build_absolute_uri(course_run_url)
-
-
-def get_emails_enabled(user, course_id):
-    """
-    Get whether or not emails are enabled in the context of a course.
-
-    Arguments:
-        user: the user object for which we want to check whether emails are enabled
-        course_id (string): the course id of the course
-
-    Returns:
-        (bool): True if emails are enabled for the course associated with course_id for the user;
-        False otherwise
-    """
-    if is_bulk_email_feature_enabled(course_id=course_id):
-        return not is_user_opted_out_for_course(user=user, course_id=course_id)
-    return None
 
 
 def get_course_run_status(course_overview, certificate_info):
